refactored Viewer2D using ViewerBase Abstract class
Refactored also the View2D using the SimpleUserInterface abstract class
